requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
decorator==5.1.1
Django==4.2.7
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.4
executing==2.0.1
ipython==8.17.2
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
prompt-toolkit==3.0.41
psycopg==3.1.13
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.13.0
typing_extensions==4.8.0
wcwidth==0.2.11

django-crispy-forms~=2.1